[sw] Port `spi_host_testutils` to devicetables
Ports the spi_host_testutils utilities to use the devicetable Pinmux
interface, so that it no longer relies on Earlgrey constants. Also
updates the corresponding tests that use this library to provide a
CSB pin through the DT interface, though the remainder of these
tests are not yet ported to devicetables.

This library uses pinmux mappings that are quite specific to
Earlgrey/Darjeeling, so relevant conditional compilation is introduced
to ensure the testutils only build information for relevant tops, and
error in the case of an undefined top (to make tracking down errors
easier).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>