[dv, aon_timer] simultaneous interrupt prediction fix
It's possible for an interrupt value to be modified via:
 - count > threshold
 - intr_state/intr_test write

If there's an interrupt due to count > threshold, and by the time it
synchronises to the sys clock happens to be a write to intr_state
clearing the interrupt, then the intr_state write wins.
Due to the above, the TB needs to remove the `count > threshold` prediction
out of the timed queue to avoid any false mismatch.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>