[dv, aon_timer] Split intr_state prediction on intr_test write
To avoid wrongly updating an intr_state bit on intr_test write the TB
predicts each field separetely depending on which bits were set on the intr_test
write.

This avoids the mismatch that could happen if the TB was going to predict an
intr_state field due to count > threshold and has set intr_status_exp but not
yet predicted the value by the time an intr_test occurs.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>