[rom_ctrl,dv] Slightly tweak an exclusion annotation
The excluded line is a continuous assignment. If the right hand side
happens to be constant then it will never be marked as executing (and
we don't care!)

I wrongly suggested that we should have an assertion saying that
it *was* constant. But that's wrong, and was removed by a920ad246c.
Tweak the text to avoid things being confusing next time.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>