[topgen,sw] Port autogen SW tests to Darjeeling
This commit ports the autogenerated `plic_all_irqs_test` and
`alert_test` software test targets so that they are generated and can be
built for the Darjeeling top, like they can already for the Earlgrey
top. This is achieved by fixing the test templates to make them
no longer Earlgrey-specific, ensuring all address spaces memory mappings
are included, and switching to using the autogenerated DIF headers
rather than the regular DIF headers (as the autogenerated headers are a
subset, sufficient for these tests, and allow us to build tests for
Darjeeling where DIFs are still missing).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>