[hsmtool] Use `strum` and `clap::ValueEnum` in `hsmtool::util:wrap`
This change adds `strum` and `clak::ValueEnum` to the
`hsmtool::util::wrap:Wrap` enum to simplify cli string formatting.

This change also switches the default import/export wrapping mechanism
for aes and kdf to `rsa-pkcs`. This is to have a default configuration
that is compatible with SoftHSM (i.e. there is no current support for
rsa-pkcs-oaep wrap/unwrap operations in SoftHSM).

Signed-off-by: Miguel Osorio <miguelosorio@google.com>