[prim] Add flop without reset
From the physical perspective its better to reduce the load on the
reset network. This PR adds a new prim for a flop without reset
It can be used where data is transmitted alongside some control
signal. The control signals would use flops with reset and always
have a known value, but the data is using a flop without a reset.
Then, the data is only used for example when the valid control
signal indicates that there is actually valid data available.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>