[hw,sram_ctrl,rtl] Ability to correct single-bit errors and log error
This PR adds the ability to correct single-bit errors using
the ECC on the data. Double bit errors are only detected.

When there is a single bit error, it is corrected and logged
via the new error log, which can be connected to a RAS error
log for example. Double errors are passed to the requester
but also logged to the error log. The error correction happens
via a new piped ECC stage, which increases the latency of the
sram_ctrl adapter.

In DJ and EG, this feature is not not used. The error log
is tied off.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>