[util/vendor] Allow patches to files
Sometimes a mapping has a "from" and "to" field that is actually a file
instead of a whole directory. Example of this can be found here:
https://github.com/lowRISC/sonata-system/blob/06102a66aebdda1d42966b049409d8ddaa6cd53a/vendor/lowrisc_ip.vendor.hjson#L15
When also adding a patch directory this runs into a problem because git
apply gets supplied the "to" field as a directory without checking whether
it is a file first. This commit fixes that by checking if it is a file
before passing it along as the base directory.

Signed-off-by: Marno van der Maas <mvdmaas+git@lowrisc.org>