[sw,math] Tweak `udiv64_slow` to avoid `__lshrdi3`
The way the code is written can cause the compiler, with certain
optimization settings, to call `__lshrdi3` instead of implementing
the shifts with 32-bit operations. This commit rewrites the code
so that all shifts on 64-bit values are by one so that any reasonable
compiler will refrain from calling `__lshrdi3`.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>