pyproject: Export python modules in util
We need to share some existing Python tools that will be used from other
projects. In order to do so, we rely on setuptools to export these
Python tools as Python modules.

Given there is a dependency from inside the topgen module to
version_file.py that is not exposed as a Python module, it isn't
available when topgen is imported from outside, simply because the
external project has no knowledge of `version_file`. To solve this
issue, I've moved version_file.py to its own module under the util
directory. This allows topgen and other modules to refer to version_file
as a known dependency to any external project that imports topgen in the
first place.

Signed-off-by: Sebastien Boeuf <seb@rivosinc.com>