[sw] Port pwrmgr_*_sleep_all_wake_ups to multitop
This does not make the test run on any other top than earlgrey yet
but everything is made more modular and it uses the DT so that no
references to earlgrey are necessary. These tests needs to be ported
all at once because they all rely on `pwrmgr_sleep_all_wake_ups_impl`
which exposes a lot of implementation details.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>