[sw,dice] Lockdown the attestation seed page to readonly
The attestation seed page contains CSPRNG generated seeds, which will be mixed
to the derived seed for generating the keypair. It is crucial that the page
should not be tampered by less secure boot stages, which could lead to key
leakage.

To prevent seed tampering, we lockdown the page to readonly at the IMM_ROM_EXT
stage which owns the first key ladder state (UDS).

Change-Id: I049d26ee85f3945b2020f7c2ab39dd7d1b4df900
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>