[bazel,otp] remove im_ext hash fusion into MANUF_STATE
With the fallback mechanism implemented in #25984, ROM can now attempt
the alternative ROM_EXT slot if the immutable ROM_EXT hash is invalid.
Consequently, the fusion of this hash into creator's `MANUF_STAGE` OTP
field and the binding of ROM_EXT to `creator_manuf_state` are no longer
needed. This removes the corresponding functionality from the OTP
tooling and the manifest update tool.

Signed-off-by: Anthony Chen <antchen@google.com>