[sw,imm_rom_ext] Transfer DICE public keys to mutable ROM_EXT.
The previous PR #26409 transferred the public key ID to mutable ROM_EXT
to validate the certificate. This is sufficient for the x509 variant,
but not for the CWT variant, which uses public key contents to validate
the UDS COSE Key. The manual e2e provisioning test
//sw/host/provisioning/orchestrator/tests:e2e_emulation_dice_cwt_cw340_test
can catch this error.

This commit fixes the test by putting the public key contents into the static
dice region with the public key ID.

Change-Id: Ic7b68869ce753bfd8f7505ac0570b59418065984
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>