[sw,e2e] Fix ROM_EXT anchor text to exclude IMM_ROM_EXT
The e2e provisioning test currently searches for the "ROM_EXT" message from the
dut console to report its version.

However, this string also matches the "IMM_ROM_EXT" startup message, which is
displayed prior to the ROM_EXT message and can cause the test to report an
incorrect version.

Change-Id: I97509c9efc34682c682121ed034d9cb53715ecac
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>