[hsmtool] Check sensitive attribute before reading AttributeType::Value
The Luna HSM PCKS#11 interface does not mark the CKA_VALUE attribute as
AttributeInfo::Sensitive even if the object has the CKA_SENSITIVE
attribute set to true. This results in the HSM throwing and error when
attempting to read the value in the `AttributeMap::from_object` method.

The solution is to explicitly check the CKA_SENSITIVE value before
attempting to read CKA_VALUE.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>