Update the Issuer/Subject name size
There are a few cases where the DICE needs to generate an identifier for
use in certificates. To ensure these identifiers are deterministic and
require no additional DICE inputs, the identifiers are derived from the
associated public key. The identifiers are 20 octets so they fit in the
RFC 5280 serialNumber field constraints and the X520SerialNumber type
when hex encoded.
ref: https://pigweed.googlesource.com/open-dice/+/HEAD/docs/specification.md#deriving-identifiers

Signed-off-by: Tommy Chiu <tommychiu@google.com>