[aon_timer,doc] Remove 32-bit wakeup timer ref in docs
When the aon timer was changed from 32 bits to 64 bits, the docs were
mostly updated (in the theory of operation / registers / programmer's
guide), but the index README page was missed in the list of features.
This is the first thing most people will read about the IP, so it should
probably be accurate.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>