[dv, aon_timer] Alternating enable ON/OFF
This VSEQ configures the counts a few cycles away from the threshold and aims to
disable the counts shortly before/after the threshold.
It ensures the TB is resilient against non SW-like stimulus by making sure the
predictions for interrupts can work correctly when abruptly disabling the count
near the threshold.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>