[dice,cwt] Measure Owner SW domain in CDI 1 mode field
Our current DICE chain has no field that can easily distinguish between device
booting with prod and dev Owner SW signing keys, since it’s implicitly measured
in the code hash. According to Open Profile for DICE, the mode field is the
canonical way to identify prod/dev/etc, and it should be set to non-Normal if
the device is not in a secure state.

Change-Id: I307474674ababd3ca27dee696b5baf5f3e426b42
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>