[sw,dice] Simplify x509 serial number checker
This PR implements a simplification similar to #26293 for X509 variant. It both
reduces size and fixes a boundary checking issue in the previous implementation.

As this check is exclusively used for cache refresh, an incorrect result doesn't
compromise security, but only affects boot time. Therefore, there is no need to
harden the check.

Change-Id: I0bf5fe85800e0bfd345cc98bc2fc52f93186b6e3
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>