[lint,soc_dbg_ctrl] Linting requires pwr_boot_status_t
In the absence of full support for virtual cores the linting of
soc_dbg_ctrl requires that the Darjeeling pwrmgr_pkg be specified
directly as a dependency because otherwise the 'pwr_boot_status_t'
structure may be unavailable since it is not present in the other
top-level specializations.
Resolving this more cleanly requires the work in PR #23555.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>