[rules] Make the default image kind configurable by exec_env
Until now, the default kind of images created by opentitan_binary
or opentitan_test was `flash`. With multitop this is inconvenient
because some execution environments may not even have something
that looks like a flash. This commit changes this behaviour: if
the `kind` attribute is not specified, the default value will now
come from the execution environment, via a new `default_kind`
attribute. For compatibility, if `default_kind` is not provided,
it defaults to `flash` which emulates the old behaviour.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>