[rules,darjeeling] Introduce a new `ctn` kind of images
Darjeeling does not have a flash controller and provides the
firmware image in the CTN address space. In particular, the
image does not need to be scrambled. This means that those
firmware cannot be built quite the same way as flash images.
A new `ctn` kind of introduced so that DV can transform the
binary to a format suitable for Darjeeling DV. The `sim_dv`
execution environment is set of `ctn` by default on Darjeeling.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>