[tl,dv] Fix infinite loop in tl_host_driver
This was my fault and I triggered the problem with commit 09af19f.
What I hadn't realised was how the flow worked when reset got
asserted. Before my previous change:

  - Suppose we're in the middle of send_a_request_body
  - The reset is asserted and the DV_SPINWAIT_EXIT macro exits.
  - We then go around the loop again, skipping the first wait.
  - Then we set req_abort=1
  - The next wait doesn't do anything (because we're in reset)
  - We get back to the top of the loop, which exits because req_abort
    is true.

This is a bit mad! It would probably make more sense to drop out on
the third bullet above, and that's what this commit does.

The reason my previous change broke things is that it switched the
last argument of send_a_request_body to be an output argument. The
effect is that we write req_abort=0 on bullet 5, so the loop doesn't
exit any more. Oh no!

Making the loop work more explicitly (matching the check for abort or
reset on line 237) seems like a good idea, and it also fixes a stupid
bug.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>