[toolchain] allow zero-sized arrays
This is a GNU extension, which we disallow by default. However in many
cases special-casing 0 makes the code more verbose without any gain, so
just enable this extension.

The behaviour of this extension is well-specified and consistent across
different Unix compilers, and is already used by a lot of projects, so
is safe to enable.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>