[tl,rtl] Add UsesDs parameter to prim_subreg and set to 1
This parameter means "the place that the prim_subreg is instantiated
cares about the output of the ds port". This isn't always true, which
eventually becomes visible in FPV coverage which points out that in
some cases, the ternary to drive ds doesn't actually have any effect
in the design:

  assign ds = wr_en ? wr_data : qs;

This commit adds a UsesDs parameter to the block to represent whether
the instantiation site actually cares about the output.

Because we check in lots of auto-generated code, the commit looks
absolutely enormous! To avoid it being impossible to review, this is
structured carefully to have no impact on design behaviour.

The only changed files that aren't auto-generated are:

  util/reggen/reg_top.sv.tpl
  hw/ip/prim/rtl/prim_subreg.sv
  hw/ip/prim/rtl/prim_subreg_shadow.sv

The change in prim_subreg.sv gets rid of the ternary line if the
parameter is not true (in order to get rid of the FPV coverage hole).
To make any bugs easier to track down, this intentionally doesn't
drive the ds output in that case. The idea is that if we accidentally
attach something, we'll get x values propagating through the design.

The instantiation inside prim_subreg_shadow gets connected correctly,
passing the information that was given to prim_subreg_shadow about
whether the instantiation of the larger module is wired up.

The change in reg_top.sv.tpl doesn't do anything: it passes the
default value of the parameter to each instantiation.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>