[tl,rtl] Set UsesDs to zero sometimes when ds is not used
The parameter was added in the previous commit and this commit
actually sets it to zero some of the time, chosen to be the
instantiations where the ds output port is not connected to anything.

Again, this commit is quite large because of auto-generated code.
There are actually only 2 lines changed in handwritten code, which are
in

  util/reggen/reg_top.sv.tpl

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>