[prim,rtl] Avoid using inside in a constant expression
The SystemVerilog spec doesn't technically allow this because the
stuff in this test is a constant expression ("constant_expression" in
section A8.3 of the spec). But this doesn't allow an
"inside_expression", unlike a normal expression.

How annoying! Expand the check to be explicit.

This was found by the Verissimo linting tool (thanks, guys!)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>