[testutils,clkmgr] Slightly delay sampling of clk_hints_status
Sampling the CLK_HINTS_STATUS register of the clkmgr after writing
to CLK_HINTS risks sampling prematurely because of the clock domain
crossings that occur.

For Earl Grey the clkmgr register interfaces operate on a clock
that is circa 4 times slower than the software-hintable clocks and
this led to test failures of chip_sw_aes_idle for some seeds (clock
ratios and phase relationships) in DV simulation.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>