[dice] Eliminate `.data` item in certificate checking
We want to avoid having a `.data` section in ROM and ROM_EXT code.

The dice cert serial number check was using a staticly initialized data
item to keep the ASN.1 prefix around.  Manually initializing this prefix
in code requires fewer bytes than the static initializer and doesn't
require the initializer to be in `.data` or `.rodata`.

Signed-off-by: Chris Frantz <cfrantz@google.com>