[tl,dv] Simplify tl_seq_item::is_ok
This was inspired by a lint warning from Verissimo about casting the
return of the function to void. That was done because the function had
a side-effect and could generate an error.

But there aren't many call sites (two!), so we can easily move that
behaviour to the call site, and this allows us to simplify/shorten
things as well.

(The diffstat on the commit doesn't look so great, but that's because
we have three templated copies of the +5/-1 change in
flash_ctrl_scoreboard.sv)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>