[rv_plic,rtl] De-duplicate rv_plic_component
Before this commit, that part of rv_plic got duplicated for each
top-level. This was a little mad: downstream users would instantiate
lots of copies of identical modules, but giving them different names
each time.

Tweak things so that we only duplicate the stuff that's genuinely
different each time. Conveniently, this is quite easy in the core
files because there was already a split into
rv_plic_component.core (since the beginning of history in the
repository).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>