[rules] Make sim_dv create a 32-bit word VMEM
The opentitan_{binary,test} rules were really thought for the case
where the firmware is stored in flash. As a result, the `kind`
argument is either `rom`, `ram` or `flash`, with the default value
being `flash`. This creates a problem: for earlgrey we need to default
to `flash` but for Darjeeling, the firmware is not stored in flash!
There are several options to fix this:

- Create a new kind (`ctn`?) for DJ and manually annotate every test
  to set the kind depending on the top: clearly not scalable.
- Create a new kind (`ctn`) for DJ and let each exec_env specify its
  default kind (instead of it being set by the rule): this is cleanest
  solution but it requires quite a bit of infrastructure change in the
  rules for a very limited use case.
- Keep with existing `flash` kind but interpret it more liberaly:
  since we don't want to rename it (churn for no reason), we think of
  `flash` as meaning "the firmware storage" which on Darjeeling DV
  means... a RAM in the CTN!

This commit follows the last approach. This poses a small difficult
however: the sim_dv exec_env code is shared between all tops and
currently only creates 64-bit word VMEM as well as scrambled VMEM.
But DJ DV wants a 32-bit word VMEM for its RAM. Since DV is run by
DVsim which simply picks the files it wants, there is no downside
to building more formats than we need. This is exactly what this commit
does: add a 32-bit word VMEM file to the output.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>