[rv_plic, fpv] Interrupts are level triggered
Since le_i is tied to as parameter which is 0, there is a deadcode
statement in rv_plic_gateway which is a ternary statement and the
conditional part is le_i. That statement is waived in
coverage_waivers.tcl. To support that waiver, an assertion is
added which checks that le_i is always false.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>