[dv] Make dv_base_env_cfg::create_ral_models safe to call twice
This will no longer create a model for a register block that exists
already.

This function gets called by initialize() and the idea is that an env
config can now set up RAL models before the call to initialize (so
that they can be passed in to the constructor, which needn't properly
know the types).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>