[hsmtool] Generalize handling of sensitive attributes in attrs.rs.
Besides `CKA_VALUE` There are aother attributes that may be sensitive
and should not be accessed if `CKA_SENSITIVE` is set to true.

This change adds the `AttributeMap::sensitive_attrs` static function to
get a manually compiled list of sensitive attributes.

The `AttributeMap:from_object` method was refactored to properly handle
potentially sensitive fields.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>