[tests,pwrmgr_sleep_all_wake_ups] Change device/host protocol
Previously, the device was just printing "Woke up by source %d"
and the host harness would pick up on that to know what to do.
However this encodes the hardware wake up sources into the harness
whereas the only thing the harness does is poke a few pins to wake
up. This commit changes that so that the device pins a string (e.g.
PINMUX, ADC_CTRL, etc) so that the harness can match per IP type
instead of knowing the wakeup sources numbers.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>