[test,pwrmgr_sleep_all_wake_ups] Port to DT
This does not make the test run on any other top than earlgrey yet
but everything is made more modular and it uses the DT so that no
references to earlgrey are necessary. This commit changes the internal
logic the library: instead of listing all wakeup sources, the library
has a library of IP function so that it can handle any instance of
certain IPs. Then the test is driven by the DT-provided list of
wakeups. This makes the test multitop aware: as long as every IP
has a corresponding entry in the library, the test will be able to run.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>