[dv, alert_receiver_driver] Driver refactor for correct alert handshake
Similarly to the monitor, the driver has been refactor so it tackles either
alerts or pings.

The driver now has the capability to retract a ping if it sees an alert before
the ping is driven.

In addition, the queues holding alert and pings are cleared during reaset, just
so if the driver has sent a ping/alert before reset and the handshake is not
completed the driver won't assume the handshake needs to continue after reset.

Finally, the driver can be configured via the config object by setting
'drive_ping' to zero so it doesn't drive pings. This is particularly useful
since the alert escalation agent has a ping sequence in the run_phase which runs
forever. Said sequence may interfere with other TB checks if a ping is driven
when the checks aren't expecting any alert such as at the end of a VSEQ.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>