[dv, cip_base_vseq] Adjust base_vseq to new alert_agent behavior
The VSEQ now sets the 'drive_ping' to 0 in the post_start method. This way, if
there happens to be a periodic ping (generated in the agent) the driver can
decide not to drive it to avoid having an alert at the end. Otherwise that
would cause some fatal alerts when the VSEQ isn't expecting any.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>