[racl_ctrl,dv] Add a core to generate a templated register model
This is reasonably trivial ("run ralgen"), but encoding it for fusesoc
is a little tricky because fusesoc only allows dependencies for
filesets. Very fortunately, the fusecoc reference manual had the magic
trick:

  https://fusesoc.readthedocs.io/en/stable/user/build_system/generators.html#calling-a-generator

Use this core as a dependency of racl_ctrl_env.core to avoid the
non-templated code needing to look at an hjson file.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>