[pentest] Return rstmgr dump to host
This commit configures the reset manager in the pentest framework to
collect reset and alert reasons. By introducing a new pentest command
this information can be returned to the host after a reset happened.

Note that the reset and alert reason is not saved when performing a
POR.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
(cherry picked from commit 4b273e0)